+++ packages/@pollyjs/cli/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.1"></a>
+## [0.2.1](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/cli/compare/@pollyjs/cli@0.2.0...@pollyjs/cli@0.2.1) (2018-06-29)
+
+
+
+
+**Note:** Version bump only for package @pollyjs/cli
+
 <a name="0.2.0"></a>
 # [0.2.0](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/cli/compare/@pollyjs/cli@0.1.1...@pollyjs/cli@0.2.0) (2018-06-27)