+++ packages/@pollyjs/node-server/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.4.0"></a>
+# [0.4.0](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/node-server/compare/@pollyjs/node-server@0.3.0...@pollyjs/node-server@0.4.0) (2018-06-29)
+
+
+### Bug Fixes
+
+* Do not display node server listening banner in quiet mode ([1be57a7](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/node-server/commit/1be57a7))
+
+
+### Features
+
+* **node-server:** Add cors support to express server to pass-through all requests ([223ce4e](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/node-server/commit/223ce4e))
+
+
+
+
 <a name="0.3.0"></a>
 # [0.3.0](https://github.com/netflix/pollyjs/tree/master/packages/@pollyjs/node-server/compare/@pollyjs/node-server@0.2.0...@pollyjs/node-server@0.3.0) (2018-06-27)